Added SoftDelete and data validation for enteties and ChatService

+++ DiscordClone.API/Controllers/ChatController.cs
@@ -1,8 +1,12 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using DiscordClone.Services.Data.Interfaces;
 using DiscordClone.ViewModels;
+using DiscordClone.ViewModels.ChatRoom;
+using System.Security.Claims;
+using Microsoft.AspNetCore.Authorization;
 namespace DiscordClone.API.Controllers
 {
+    [Authorize]
     [ApiController]
     [Route("api/chat")]
     public class ChatController : Controller
@@ -32,24 +36,54 @@ namespace DiscordClone.API.Controllers
         }
         [HttpPost]
         [Route("")]
-        public async Task<IActionResult> CreateChat(ChatRoomViewModel chat)
+        public async Task<IActionResult> CreateChat(ChatRoomInputModel chat)
         {
-            Guid chatId = await _chatService.CreateChatAsync(chat);
-            chat.Id = chatId;
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if (userId == null)
+            {
+                return Unauthorized();
+            }
+            Guid chatId = await _chatService.CreateChatAsync(chat, Guid.Parse(userId));
             return CreatedAtAction(nameof(GetChat), new { id = chatId.ToString() }, chat);
+        }
+        [HttpPost]
+        [Route("update/{id}")]
+        public async Task<IActionResult> UpdateChat(string id, ChatRoomViewModel chat)
+        {
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if (userId == null)
+            {
+                return Unauthorized();
+            }
+            var result = await _chatService.UpdateChatAsync(chat, Guid.Parse(userId));
+            if (!result)
+            {
+                return NotFound();
+            }
+            return Ok();
         }
         [HttpPost]
         [Route("message")]
         public async Task<IActionResult> PostMessage(MessageViewModel message)
         {
-            var messageId = await _chatService.PostMessageAsync(message);
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if (userId == null)
+            {
+                return Unauthorized();
+            }
+            var messageId = await _chatService.PostMessageAsync(message, Guid.Parse(userId));
             return CreatedAtAction(nameof(PostMessage), new { id = messageId }, message);
         }
         [HttpDelete]
         [Route("{id}")]
         public async Task<IActionResult> DeleteChat(string id)
         {
-            var result = await _chatService.DeleteChatAsync(Guid.Parse(id));
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if (userId == null)
+            {
+                return Unauthorized();
+            }
+            var result = await _chatService.SoftDeleteChatAsync(Guid.Parse(id), Guid.Parse(userId));
             if (!result)
             {
                 return NotFound();
@@ -60,7 +94,12 @@ namespace DiscordClone.API.Controllers
         [Route("message/{id}")]
         public async Task<IActionResult> DeleteMessage(string id)
         {
-            var result = await _chatService.DeleteMessageAsync(Guid.Parse(id));
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if (userId == null)
+            {
+                return Unauthorized();
+            }
+            var result = await _chatService.SoftDeleteMessageAsync(Guid.Parse(id), Guid.Parse(userId));
             if (!result)
             {
                 return NotFound();
@@ -69,8 +108,13 @@ namespace DiscordClone.API.Controllers
         }
         [HttpPost]
         [Route("{id}/users/{userId}")]
-        public async Task<IActionResult> AddUserToChat(string id, string userId)
+        public async Task<IActionResult> AddUserToChat(string id)
         {
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if (userId == null)
+            {
+                return Unauthorized();
+            }
             var result = await _chatService.AddUserToChatAsync(Guid.Parse(id), Guid.Parse(userId));
             if (!result)
             {
@@ -80,8 +124,13 @@ namespace DiscordClone.API.Controllers
         }
         [HttpDelete]
         [Route("{id}/users/{userId}")]
-        public async Task<IActionResult> RemoveUserFromChat(string id, string userId)
+        public async Task<IActionResult> RemoveUserFromChat(string id)
         {
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if (userId == null)
+            {
+                return Unauthorized();
+            }
             var result = await _chatService.RemoveUserFromChatAsync(Guid.Parse(id), Guid.Parse(userId));
             if (!result)
             {
